Remove automatic presence-checking
In VectorDraw, the grader first automatically checks for every vector that appears in `expected_results`. This is convenient for many problems, but for FBDs we would like to specifically check for the absence of "fake" forces.

Note that this means `expected_results` must explicitly contain presence checks! Otherwise edX will give python errors.